Create __init__.py
This file serves as an **ontological marker**. Its sole purpose is to tell Python that the directory it resides in is a package. This allows you to import modules from this directory in other parts of your project and in projects by other developers. Without this file, Python will not recognize your directory as a package.